pyproject.toml
[project]
name = "wrecking-crew"
version = "0.1.0"
description = "A catalog of archetype critics for pressure-testing your work in any agentic coding tool."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
  "PyYAML>=6.0",
  "sentence-transformers>=2.2",
  "sqlite-vec>=0.1",
  "numpy>=1.24",
  "mcp>=1.0",
]

[project.scripts]
crew = "wrecking_crew.cli:main"
crew-mcp = "wrecking_crew.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wrecking_crew"]
